requirements.txt
langchain>=0.2.0
langchain-postgres>=0.0.8
langchain-community>=0.2.0
streamlit>=1.31.0
psycopg>=3.2.9
psycopg-binary>=3.2.9
psycopg2>=2.9.5
psycopg2-binary>=2.9.5
python-dotenv>=1.0.1
requests>=2.31.0
jupyter>=1.0.0
pytest>=8.0.0
sqlalchemy>=2.0.0